requirements.txt
Django==5.0.1
djangorestframework==3.14.0
gunicorn==21.2.0
psycopg[binary]>=3.2.0
dj-database-url==2.1.0
whitenoise==6.6.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
pyjwt==2.8.0